Modify files which make a test fail exactly once
Changed just the name of a test in `peer-network.test.ts`, and extended the
minimum timeout of the test node script in `package.json`.  Running the
modified test results in a much longer runtime along with test failure, but
only once.  Running the test a second time results in test success. This
could be an indication that some kind of caching is producing incorrect test
results.