Add test to confirm network drops transactions with negative fee
Added a test which verifies that a node, when sent a transaction with a
negative fee, does not add the transaction to the mempool, and does not
forward the transaction to other peers.  Actually constructing a posted
transaction with a negative fee is not possible with the current native Rust
implementations, so small modifications were necessary in `ironfish-rust` and
`ironfish-rust-nodejs` in order to produce a transaction of the proper form
for this test.