Add tests to confirm replay issue for "pure mint" transactions
Tests verify that it is possible for a posted transaction spending no notes to
be accepted by the blockchain multiple times.  This is a niche situation, as
such a transaction can only have mint inputs and cannot have a positive
transaction fee.  However, a "pure mint" transaction which mints a custom
asset and produces a corresponding output note is an example.

A first test confirms that a pure mint transaction received from the network
is accepted by a node as valid, and is forwarded to other connected peers.

A second test confirms that a pure mint transaction in a node's memory pool
will be added to mining block templates and submitted to a miner via RPC.

A third test confirms that a single posted pure mint transaction can be
included in multiple blocks on the blockchain under currently implemented
consensus rules.