Add tests to verify mempool contamination by inline double-spends
Tests verify that a transaction with an "inline double-spend", i.e. containing
more than one spend with the same nullifier, can contaminate the mempool of a
node connected to the network, potentially restricting node functionality
without manual intervention.

A first test confirms that an inline double-spend transaction received from the
network is accepted by a node as valid, is added to the node's mempool, and is
forwarded to connected peers.

A second test confirms that an inline double-spend in a node's mempool will be
added to mining block templates constructed by the node's mining manager.

A third test confirms that a mining block template containing an inline
double-spend is NOT submitted to miners by RPC.  The mining manager first
attempts to add the block template to the current blockchain head, which runs
the true network consensus verification on the block template.  This results
in rejection of the block containing the inline double-spend transaction
before any mining is done.